Moving quota classes to a separate package
