+++ vimeo-networking/src/main/java/com/vimeo/networking/model/uploadquota/Space.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package com.vimeo.networking.model;
+package com.vimeo.networking.model.uploadquota;
 
 import com.google.gson.annotations.SerializedName;
 import com.vimeo.stag.UseStag;
